Fix race condition resulting in File sharing violation
This race condition is a bit complex to explain - basically, you trigger
it by requesting the same key on two different threads at the same time.
When this happens, we kick off what *appears* to be an async file open
(via Utility.SafeFileOpenAsync).

However, file opens in NT are always synchronous - if you're not
explicitly deferring them via a thread pool item, you're executing them
in-context. In this case, executing the ".Multicast().Connect()" was the
line that actually ended up sending the call down to CreateFile.

Because MemoizedMRUCache was still adding the key at the time (via
CalculationFunction) and stuck in the CreateFile call, the 2nd thread
experiences the cache miss. If you get unlucky and thread #2 makes enough
progress, it gets a file sharing violation.