Hacky first step fix for vacuum
You can't have multiple statements in a prepared statement so the vacuum
command was never been executed. This splits it into two statements that
are run sequentially. Problem number two is that the vacuum command
can't be run in a transaction[1] so we commit the transaction right
before the vacuum and then start it up again after. This will very
likely have unintended side-effects but it's a proof that this is what's
going on.

1. https://sqlite.org/lang_vacuum.html