Make sure we don't deadlock in Vacuum
There's a race condition in Vacuum where in it attempts to pre-empt the
main queue by scheduling a DoNothing operation but does so before
waiting on the lock and the lock itself has no guarantees on ordering.
To demonstrate, add a delay between the scheduling and the lock wait.

This. Fixes. That. (hopefully)