Simplify flush
@paulcbetts This is the thing we discussed in chat. I believe that even
with the coalescing this version will hold up to the guarantee that

"Flush ​_should_​ guarantee that, "All operations ​_issued_​ before the
Flush was called will be processed by the time Flush returns"

It can do so because while MarshalCompletion signals on the AsyncSubject
as soon as it has run the expression it will use the multicast of the
commitSubject to signal completion. So when the CompletionAsUnit
completes you can be sure that every operation queued up until the point
Flush was called has been performed.