Allow UTC DateTimes to be forced.
JSON.NET's BSON infrastructure writes DateTimes as UTC and reads them back as Local. There is a configuration property for this, but it exists only on the reader and writer - JsonSerializerSettings does not affect it. This PR adds the ability for client code to force a particular setting on JsonReader.DateTimeKindHandling.